pyproject.toml
[tool.poetry]
name = "photoarchive"
version = "0.1.0"
description = ""
authors = ["John W. Terrell <john@coolpeoplenetworks.com>"]

[tool.poetry.dependencies]
python = "^3.9"
ExifRead = "^2.3.2"
CouchDB = "^1.2"
PyExifTool = "^0.4.11"
uuid = "^1.30"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
paimport = "photoarchive.import:main"
